requirements.txt
# Frappe framework
frappe>=15.0.0

# Google Generative AI SDK
google-generativeai==0.8.5

# JWT support
PyJWT==2.3.0

mysqlclient>=2.2.0